fix(commenting): resolve ambiguous edits away from text-range anchors
Moves text-range anchor tracking into the commenting package, where it is
covered by CI, and fixes a boundary bug found while reviewing it.

An edit is often ambiguous: typing "c" before "comment" gives "ccomment",
and nothing in the two strings says which "c" is new. The diff resolved that
by maximising the common prefix, which attributed the insertion after the
anchor's start and let the comment stretch over the character just typed
against it.

Ambiguous edits now get mapped both ways and the readings are intersected,
so the result is never wider than either one and an ambiguous edit always
resolves away from the comment.

The property test previously asserted its invariant against the edit the
diff inferred rather than the edit actually applied, which is what let this
through; it now checks the applied edit.
